fastboot: Updated sparse API with 64-bit length
To support large file, the length data type should be updated
to int64_t from int.

Reviewed-on: http://gerrit.mot.com/687095
Tested-by: Jira Key <jirakey@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>
SLTApproved: Connie Zhao <czhao1@motorola.com>

+++ fastboot/protocol.c
@@ -220,7 +220,7 @@ int fb_download_data(usb_handle *usb, const void *data, unsigned size)
 static char usb_buf[USB_BUF_SIZE];
 static int usb_buf_len;
 
-static int fb_download_data_sparse_write(void *priv, const void *data, int len)
+static int fb_download_data_sparse_write(void *priv, const void *data, int64_t len)
 {
     int r;
     usb_handle *usb = priv;
